+++ client45/src/main/java/vo/receiptvo/DebitAndPayBillVO.java
@@ -1,12 +1,14 @@
 package vo.receiptvo;
 
+import java.math.BigDecimal;
+
 import state.ReceiptType;
 
 public class DebitAndPayBillVO  extends ReceiptVO{
-	public double money;
+	public BigDecimal money;
 	/**单据状态*/
 	
-	public DebitAndPayBillVO(String id, ReceiptType type, double money) {
+	public DebitAndPayBillVO(String id, ReceiptType type, BigDecimal money) {
 		super(id, type);
 		this.money = money;
 	}
